+++ acro/__init__.py
@@ -1,1 +1,1 @@
-from .safe_crosstab import safe_crosstab
+from .acro import ACRO
